+++ Graph/Targets.lua
@@ -376,6 +376,11 @@ function T:Legal(action, state, descriptor)
             action, state, tooltip)
     end
     if blocker then return false, blocker end
+    if XelAssist.Graph.WarlockNightfall then
+        tooltip, blocker = XelAssist.Graph.WarlockNightfall:PrepareLegal(
+            action, state, tooltip)
+    end
+    if blocker then return false, blocker end
     blocker = XelAssist.Graph.ClassMechanics
         and XelAssist.Graph.ClassMechanics:Blocker(
             action, state, descriptor, tooltip)